requirements.txt
fastapi>=0.116.1,<0.117.0
uvicorn>=0.35.0,<0.36.0
redis>=6.2.0,<7.0.0
httpx>=0.28.1,<0.29.0
pyyaml>=6.0.2,<7.0.0
opentelemetry-api>=1.35.0,<2.0.0
opentelemetry-sdk>=1.35.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.48b0
tiktoken>=0.9.0,<0.10.0
click>=8.2.1,<9.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx[testing]>=0.24.0